requirements.txt
ultralytics
streamlit
opencv-python
Pillow
ipykernel